requirements.txt
flask==2.2.3
werkzeug==2.2.3
tensorflow==2.10.0
pillow==9.2.0
numpy==1.21.5
scipy==1.7.3